+++ src/main/scala/com/atomist/rug/test/gherkin/handler/AbstractHandlerScenarioWorld.scala
@@ -11,6 +11,7 @@ import com.atomist.rug.spi.Handlers.Plan
 import com.atomist.rug.spi.TypeRegistry
 import com.atomist.rug.test.gherkin.{Definitions, ScenarioWorld}
 import com.atomist.tree.{TreeMaterializer, TreeNode}
+import com.atomist.util.lang.JavaScriptArray
 
 /**
   * Superclass for Handler worlds. Handles plan capture and exposing to JavaScript
@@ -18,12 +19,13 @@ import com.atomist.tree.{TreeMaterializer, TreeNode}
 abstract class AbstractHandlerScenarioWorld(definitions: Definitions, rugs: Option[Rugs])
   extends ScenarioWorld(definitions, rugs) {
 
-  private var planOption: Option[Plan] = None
+  private var recordedPlans = List[Plan]()
 
   protected def createRugContext(tm: TreeMaterializer): RugContext =
     new FakeRugContext("team_id", tm)
 
-  private var rootContext: SimpleContainerGraphNode = SimpleContainerGraphNode.empty("root", TreeNode.Dynamic)
+  private var rootContext: SimpleContainerGraphNode =
+    SimpleContainerGraphNode.empty("root", TreeNode.Dynamic)
 
   /**
     * Return the editor with the given name or throw an exception
@@ -50,32 +52,41 @@ abstract class AbstractHandlerScenarioWorld(definitions: Definitions, rugs: Opti
     rootContext = rootContext.addRelatedNode(gn)
   }
 
-  protected def recordPlan(plan: Option[Plan]): Unit = {
+  protected def recordPlan(plan: Plan): Unit = {
     //println(s"Recorded plan option $plan")
-    this.planOption = plan
+    recordedPlans = recordedPlans :+ plan
   }
 
   /**
-    * Return the plan or throw an exception if none was recorded
+    * Return a single plan or throw an exception if none was recorded
     */
-  def requiredPlan: jsScalaHidingProxy = {
-    //println(s"Contents of recorded plan: $planOption")
-    planOption.map(jsScalaHidingProxy.apply(_)).getOrElse(throw new IllegalArgumentException("No plan was recorded"))
+  def requiredPlan: jsScalaHidingProxy = recordedPlans match {
+    case Nil =>
+      throw new IllegalArgumentException("No plan was recorded")
+    case plan :: Nil =>
+      //println(s"Contents of recorded plan: $planOption")
+      jsScalaHidingProxy(plan)
+    case _ =>
+      throw new IllegalArgumentException(s"Expected exactly one plan but found ${recordedPlans.size}")
   }
 
   /**
     * Return the plan or null if none was recorded
     */
-  def plan: jsScalaHidingProxy = {
-    planOption.map(jsScalaHidingProxy.apply(_)).orNull
-  }
+  def plan: jsScalaHidingProxy =
+    recordedPlans.headOption.map(jsScalaHidingProxy.apply(_)).orNull
+
+  /**
+    * Return all plans recorded
+    */
+  def plans: JavaScriptArray[jsScalaHidingProxy] = JavaScriptArray.fromSeq(recordedPlans.map(jsScalaHidingProxy.apply(_)))
 
   /**
     * Return the message or null if none was recorded
     */
   def message: jsScalaHidingProxy = {
-    planOption match {
-      case Some(p) => p.messages.map(jsScalaHidingProxy.apply(_)).head
+    recordedPlans match {
+      case List(p) => p.messages.map(jsScalaHidingProxy.apply(_)).head
       case _ => null
     }
   }
@@ -84,23 +95,26 @@ abstract class AbstractHandlerScenarioWorld(definitions: Definitions, rugs: Opti
     * Is the plan internally valid? Do the referenced handlers and other operations exist?
     */
   def planIsInternallyValid(): Boolean = {
-    val p = planOption.getOrElse(throw new IllegalArgumentException("No plan was recorded"))
-    !p.instructions.map(_.instruction).exists {
-      case Edit(detail) =>
-        val knownEditors: Seq[String] = rugs.map(_.editorNames).getOrElse(Nil)
-        !knownEditors.contains(detail.name)
-      case Generate(detail) =>
-        val knownGenerators: Seq[String] = rugs.map(_.generatorNames).getOrElse(Nil)
-        !knownGenerators.contains(detail.name)
-      case Review(detail) =>
-        val knownReviewers: Seq[String] = rugs.map(_.reviewerNames).getOrElse(Nil)
-        !knownReviewers.contains(detail.name)
-      case Command(detail) =>
-        val knownCommandHandlers: Seq[String] = rugs.map(_.commandHandlerNames).getOrElse(Nil)
-        !knownCommandHandlers.contains(detail.name)
+    if (recordedPlans.isEmpty)
+      throw new IllegalArgumentException("No plan was recorded")
+    recordedPlans.forall(p => {
+      !p.instructions.map(_.instruction).exists {
+        case Edit(detail) =>
+          val knownEditors: Seq[String] = rugs.map(_.editorNames).getOrElse(Nil)
+          !knownEditors.contains(detail.name)
+        case Generate(detail) =>
+          val knownGenerators: Seq[String] = rugs.map(_.generatorNames).getOrElse(Nil)
+          !knownGenerators.contains(detail.name)
+        case Review(detail) =>
+          val knownReviewers: Seq[String] = rugs.map(_.reviewerNames).getOrElse(Nil)
+          !knownReviewers.contains(detail.name)
+        case Command(detail) =>
+          val knownCommandHandlers: Seq[String] = rugs.map(_.commandHandlerNames).getOrElse(Nil)
+          !knownCommandHandlers.contains(detail.name)
         // TODO there are probably more cases here
-      case _ => false
-    }
+        case _ => false
+      }
+    })
   }
 
   private class FakeRugContext(val teamId: String, _treeMaterializer: TreeMaterializer) extends RugContext {